refactor : final 적용
